# handshake: optimize AEAD handling for long header sealers and openers