Not a huge improvement, only slightly fewer allocs.

Part of #3663.

```
name                               old time/op    new time/op    delta
InitialAEADCreate-16                 4.78µs ± 5%    4.79µs ± 5%    ~     (p=0.927 n=10+10)
InitialAEAD/opening_100_bytes-16      116ns ± 2%     118ns ± 3%    ~     (p=0.116 n=8+10)
InitialAEAD/opening_1200_bytes-16     207ns ± 1%     208ns ± 3%    ~     (p=0.616 n=10+10)
InitialAEAD/sealing_100_bytes-16      111ns ± 3%     111ns ± 3%    ~     (p=0.764 n=9+10)
InitialAEAD/sealing_1200_bytes-16     206ns ± 2%     207ns ± 3%    ~     (p=0.424 n=10+10)

name                               old alloc/op   new alloc/op   delta
InitialAEADCreate-16                 10.5kB ± 0%    10.4kB ± 0%  -0.72%  (p=0.000 n=10+10)
InitialAEAD/opening_100_bytes-16      0.00B          0.00B         ~     (all equal)
InitialAEAD/opening_1200_bytes-16     0.00B          0.00B         ~     (all equal)
InitialAEAD/sealing_100_bytes-16      0.00B          0.00B         ~     (all equal)
InitialAEAD/sealing_1200_bytes-16     0.00B          0.00B         ~     (all equal)

name                               old allocs/op  new allocs/op  delta
InitialAEADCreate-16                    131 ± 0%       129 ± 0%  -1.53%  (p=0.000 n=10+10)
InitialAEAD/opening_100_bytes-16       0.00           0.00         ~     (all equal)
InitialAEAD/opening_1200_bytes-16      0.00           0.00         ~     (all equal)
InitialAEAD/sealing_100_bytes-16       0.00           0.00         ~     (all equal)
InitialAEAD/sealing_1200_bytes-16      0.00           0.00         ~     (all equal)